requirements.txt
rich
PyYAML
openai
requests